toolkits/llm_rag/requirements.txt
# LLM-RAG Toolkit Requirements
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.11.0
pyyaml>=6.0
click>=8.1.0
pandas>=1.5.0
requests>=2.28.0
numpy>=1.23.0
